+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `counted_float benchmark` can write results to a JSON file via `--output`
+
 ### Changed
 
 ### Deprecated